% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.pkg.R
\docType{package}
\name{example.pkg.Rcpp.splines2}
\alias{example.pkg.Rcpp.splines2}
\title{Example Package using splines2 with Rcpp}
\description{
An example package that demonstrats the C++ interface of the splines2
package with the help of Rcpp.
}
